requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
distlib==0.3.4
Django==3.1.3
Django-Basic-CRUD-Generator==0.2
django-bootstrap==0.2.4
django-bootstrap-pagination==1.7.1
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-filter==21.1
django-sslserver==0.22
djangorestframework==3.13.1
filelock==3.6.0
numpy==1.22.3
pandas==1.4.2
Pillow==8.0.1
platformdirs==2.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
